Вынес функции отрисовки элементов в модуль render.js, написал класс контроллера TripController и вынес в него логику отрисовки и навешивания обработчиков
